Rename vyos2 over vyos
